lidar/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
addict=2.3.0=py38h32f6830_2
alphashape=1.3.1=pyh44b312d_0
argon2-cffi=20.1.0=py38h27cfd23_1
async_generator=1.10=pyhd3eb1b0_0
attrs=20.3.0=pyhd3deb0d_0
autopep8=1.5.6=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
boost-cpp=1.72.0=h9d3c048_4
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2020.12.5=ha878542_0
cairo=1.16.0=hf32fb01_1
certifi=2020.12.5=py38h578d9bd_1
cffi=1.14.5=py38h261ae71_0
cfitsio=3.470=h2e3daa1_7
chardet=3.0.4=py38h06a4308_1003
click=7.1.2=pyh9f0ad1d_0
click-log=0.3.2=pyh9f0ad1d_0
click-plugins=1.1.1=py_0
cligj=0.7.1=pyhd8ed1ab_0
cloudpickle=1.6.0=py_0
cryptography=3.4.7=py38hd23ed53_0
curl=7.76.1=h979ede3_1
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38h7b6447c_0
dask-core=2021.4.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=5.0.6=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
eigen=3.3.9=h4bd325d_1
entrypoints=0.3=py38_0
expat=2.3.0=h2531618_2
ffmpeg=4.3.1=hca11adc_2
fiona=1.8.19=py38hdb5a769_0
flann=1.9.1=h2e58136_1008
fontconfig=2.13.1=hba837de_1005
freetype=2.10.4=h5ab3b9f_0
freexl=1.0.6=h7f98852_0
fsspec=0.9.0=pyhd3eb1b0_0
gdal=3.2.2=py38h507a4fd_3
geojson=2.5.0=py_0
geopandas=0.9.0=pyhd8ed1ab_0
geos=3.9.1=h9c3ff4c_2
geotiff=1.6.0=hcf90da6_5
gettext=0.21.0=hf68c758_0
giflib=5.2.1=h36c2ea0_2
glib=2.68.0=h9c3ff4c_2
glib-tools=2.68.0=h9c3ff4c_2
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h23475e2_0
gst-plugins-base=1.18.4=h29181c9_0
gstreamer=1.18.4=h76c114f_0
harfbuzz=2.8.0=h83ec7ef_1
hdf4=4.2.13=h10796ff_1005
hdf5=1.10.6=nompi_h7c3c948_1111
icu=68.1=h2531618_0
idna=2.10=pyhd3eb1b0_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py38h06a4308_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2021.2.0=h06a4308_610
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.22.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jasper=1.900.1=hd497a04_4
jedi=0.17.0=py38_0
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h36c2ea0_0
json-c=0.15=h98cffda_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kealib=1.4.14=hcc255d8_2
kiwisolver=1.3.1=py38h2531618_0
krb5=1.17.1=h173b8e3_0
lame=3.100=h7b6447c_0
laspy=1.7.0=pyh5ca1d4c_0
lastools=20171231=h9c3ff4c_1002
laszip=3.4.3=he1b5a44_1
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=hea4e1c9_2
libblas=3.9.0=1_h6e990d7_netlib
libcblas=3.9.0=3_h893e4fe_netlib
libclang=11.1.0=default_ha53f305_0
libcurl=7.76.1=hc4aaa36_1
libdap4=3.20.6=hd7c4107_2
libedit=3.1.20210216=h27cfd23_1
libev=4.33=h7b6447c_0
libevent=2.1.10=hcdb4288_3
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h2828fa1_18
libgdal=3.2.2=hbf32332_3
libgfortran-ng=7.5.0=h14aa051_19
libgfortran4=7.5.0=h14aa051_19
libglib=2.68.0=h3e27bee_2
libgomp=9.3.0=h2828fa1_18
libiconv=1.16=h516909a_0
libidn2=2.3.0=h27cfd23_0
libkml=1.3.0=hd79254b_1012
liblapack=3.9.0=3_h893e4fe_netlib
liblapacke=3.9.0=3_h893e4fe_netlib
libllvm11=11.1.0=hf817b99_0
libnetcdf=4.8.0=nompi_hfa85936_101
libnghttp2=1.43.0=h812cca2_0
libopencv=4.5.1=py38hfb05350_1
libpng=1.6.37=hbc83047_0
libpq=13.2=hfd2b0eb_2
libprotobuf=3.15.8=h780b84a_0
librttopo=1.1.0=h1185371_6
libsodium=1.0.18=h7b6447c_0
libspatialindex=1.9.3=h9c3ff4c_3
libspatialite=5.0.1=h20cb978_4
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.3.0=h6de172a_18
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=hdc55705_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.10=h72842e0_3
libzip=1.7.3=h4de3113_0
locket=0.2.1=py38h06a4308_1
lz4-c=1.9.3=h2531618_0
markdown-it-py=0.6.2=pypi_0
markupsafe=1.1.1=py38h7b6447c_0
matplotlib-base=3.3.4=py38h62a2d02_0
mdit-py-plugins=0.2.6=pypi_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.3.0=py38h54f3939_0
mkl_random=1.1.1=py38h0573a6f_0
more-itertools=8.7.0=pypi_0
munch=2.5.0=py_0
mysql-common=8.0.23=ha770c72_1
mysql-libs=8.0.23=h935591d_1
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py38_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.2=hbbd107a_1
networkx=2.5=py_0
notebook=6.3.0=py38h06a4308_0
nspr=4.30=h9c3ff4c_0
nss=3.63=hb5efdd6_0
numpy=1.20.1=pypi_0
numpy-base=1.19.2=py38hfa32c7d_0
olefile=0.46=py_0
open3d=0.11.2=py38_0
opencv=4.5.1=py38h578d9bd_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hf7af979_0
openssl=1.1.1k=h7f98852_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.4=py38h1abd341_0
pandoc=2.12=h06a4308_0
pandocfilters=1.4.3=py38h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
pcl=1.9.1=h2dfa329_1005
pclpy=0.12.0=py38_1
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.2=py38he98fc37_0
pip=21.0.1=py38h06a4308_0
pixman=0.40.0=h7b6447c_0
plotly=4.14.3=py_0
pluggy=0.13.1=pypi_0
plyfile=0.7.3=pyh44b312d_0
poppler=21.03.0=h93df280_0
poppler-data=0.4.10=0
postgresql=13.2=h6303168_2
progress=1.5=py_1
proj=8.0.0=h277dcde_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py=1.10.0=pypi_0
py-opencv=4.5.1=py38hd0cf306_1
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pygments=2.8.1=pyhd3eb1b0_0
pyhull=2015.2.1=py38hc10631b_2
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyproj=3.0.1=py38hdf83a33_1
pyqt-impl=5.12.3=py38h7400c14_7
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.8.1=pypi_0
pyqtwebengine=5.12.1=py38h7400c14_7
pyrsistent=0.17.3=py38h7b6447c_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.8=hffdb5ce_0_cpython
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytz=2021.1=pyhd8ed1ab_0
pywavelets=1.1.1=py38h7b6447c_2
pyyaml=5.3.1=py38h7b6447c_1
pyzmq=20.0.0=py38h2531618_1
qdarkstyle=3.0.2=pyhd8ed1ab_0
qhull=2020.2=h4bd325d_1
qt=5.12.9=hda022c4_4
qtpy=1.9.0=py_0
readline=8.1=h27cfd23_0
requests=2.24.0=py_0
retrying=1.3.3=py_2
rtree=0.9.7=py38h02d302b_1
scikit-image=0.17.2=py38hdf5156a_0
scikit-learn=0.24.1=py38ha9443f7_0
scipy=1.6.2=py38h91f5cce_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py38h06a4308_0
shapely=1.7.1=py38h4fc1155_4
six=1.15.0=py38h06a4308_0
sqlite=3.35.4=hdfb4753_0
terminado=0.9.3=pypi_0
testpath=0.4.4=pyhd3eb1b0_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.3.16=pypi_0
tiledb=2.2.7=h91fcb0e_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.1=pyhd3eb1b0_0
tornado=6.1=py38h27cfd23_0
tqdm=4.59.0=pyhd8ed1ab_0
traitlets=5.0.5=pyhd3eb1b0_0
trimesh=3.9.14=pyh44b312d_0
tzcode=2021a=h7f98852_1
tzdata=2021a=he74cb21_0
urllib3=1.25.11=py_0
utm=0.6.0=pyh9f0ad1d_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38_0
x264=1!161.3030=h7f98852_0
xerces-c=3.2.3=h9d8b166_2
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
